Moved UndoAction helper classes to separate UndoActions.kt file
